:truck: move preferences creators to components/Preferences
